Temporary Bugfix To Speed Up Specimen Database Loading
This will be somewhat unnecessary after caching is impmlemented,
though, this change won't hurt.  The issue was associated with
the fact that we were processing files we didn't need to
(namely backup files and resource folders) on every single page
load.  This limits us to just .sdb and .sdg files.