Error Handling to Prevent Problems Related to Checklist Naming when Filename is Blank
This hotfix is related to a bug in the limatix-qautils; however, better error handling is needed here regardless to prevent significant unintended consequences in case of future issues and also from a security perspective.  More enhancement in this regard is likely needed in the future to further sanitize potentially unsafe user input.